chore(refactor): move `ThemeJson` and `SectionNames` to `Api` namespace, update references across codebase, and remove outdated namespace aliases
