+++ tests/_data/fixtures/input-data.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace ItalyStrap\Tests;
 
-use ItalyStrap\ThemeJsonGenerator\Domain\ThemeJson\SectionNames;
+use ItalyStrap\ThemeJsonGenerator\Api\SectionNames;
 
 return [
     SectionNames::VERSION => 1,
